requirements.txt
pandas
rapidfuzz
streamlit
openai
python-dotenv